+++ firedantic/models.py
@@ -5,7 +5,7 @@ import pydantic
 from google.cloud.firestore_v1 import CollectionReference, DocumentReference
 
 from firedantic.configurations import CONFIGURATIONS
-from firedantic.exceptions import ModelNotFoundError
+from firedantic.exceptions import CollectionNotDefined, ModelNotFoundError
 
 TModel = TypeVar("TModel", bound="Model")
 
@@ -22,7 +22,7 @@ class Model(pydantic.BaseModel, ABC):
 
     def save(self) -> None:
         """Saves this model in the database."""
-        data = self.dict()
+        data = self.dict(by_alias=True)
         if "id" in data:
             del data["id"]
 
@@ -82,8 +82,30 @@ class Model(pydantic.BaseModel, ABC):
         return cls(**data)
 
     @classmethod
+    def truncate_collection(cls, batch_size: int = 128) -> int:
+        """Removes all documents inside a collection.
+
+        :param batch_size: Batch size for listing documents.
+        :return: Number of removed documents.
+        """
+        count = 0
+        col_ref = cls._get_col_ref()
+
+        while True:
+            deleted = 0
+            for doc in col_ref.limit(batch_size).stream():
+                doc.reference.delete()
+                deleted += 1
+
+            count += deleted
+            if deleted < batch_size:
+                return count
+
+    @classmethod
     def _get_col_ref(cls) -> CollectionReference:
         """Returns the collection reference."""
+        if cls.__collection__ is None:
+            raise CollectionNotDefined(f"Missing collection name for {cls.__name__}")
         return CONFIGURATIONS["db"].collection(
             CONFIGURATIONS["prefix"] + cls.__collection__
         )